Move process stuff to internal namespace
