+++ src/Cli/Internal/DefaultProcessFactory.cs
@@ -1,6 +1,7 @@
 using System.Diagnostics;
+using Cli.Services;
 
-namespace Cli.Services
+namespace Cli.Internal
 {
     internal class DefaultProcessFactory : IProcessFactory
     {
