+++ src/Cli/Services/ProcessService.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Cli.Internal;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
